feat(ui): Handle project redirect in API client
Previously we would only be handling API responses with a 302 in `ProjectContext`, which is only a subset of requests that can return a 302. Move the handling up to the API client.

Also changes the 302 response body to be an object more consistent with our error response objects.

Fixes JAVASCRIPT-3JM