feat(identity): Don't relink identity if login depends on it
When a user sets up an integration, if their identity is linked to a different user or user is linked to a different identity, we delete those before creating a new link. Now that we're using identity for login, we restrict this deletion to users that have a password so that they don't get locked out if we delete an identity they use for login.